src/ocr_engine.py
import pytesseract
from PIL import Image
import io

class OCRProcessor:
    def extract_text(self, image_bytes: bytes) -> str:
        try:
            image = Image.open(io.BytesIO(image_bytes))
            # Convert to RGB if necessary (e.g., for PNGs with transparency)
            if image.mode != 'RGB':
                image = image.convert('RGB')
            
            text = pytesseract.image_to_string(image)
            return text.strip()
        except Exception as e:
            print(f"OCR Error: {e}")
            return ""